% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSimulationResults.R
\name{aggMax}
\alias{aggMax}
\title{Sane max for aggregation}
\usage{
aggMax(x)
}
\arguments{
\item{x}{[\code{vector(n)}] \cr
numerical or character argument}
}
\value{
Maximum value in \sQuote{x} or \sQuote{NA} if \sQuote{x} is empty.
}
\description{
Sane max for aggregation
}
\details{
R's \code{max} returns \sQuote{-Inf} for empty lists. 
This is undesirable for aggregation where we would rather have \sQuote{NA}.
}
