% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSimulationReplications.R
\name{aggregateSimulationReplications}
\alias{aggregateSimulationReplications}
\title{Summarize replications of a simulation}
\usage{
aggregateSimulationReplications(results)
}
\arguments{
\item{results}{Raw simulation results. 
The result of \code{\link{extractSimulationResults}}.}
}
\description{
Summarize replications of a simulation
}
