% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BabsimDailyCases.R
\name{autoplot.BabsimDailyCases}
\alias{autoplot.BabsimDailyCases}
\title{Visualize new daily cases}
\usage{
autoplot.BabsimDailyCases(
  object,
  xlab = "Date",
  ylab = "Cases",
  clab = "Age",
  droplevels = TRUE,
  drop.NA = TRUE,
  ...
)
}
\arguments{
\item{object}{[\code{BabsimCases}] \cr
new daily cases.}

\item{xlab}{[\code{character(1)}] \cr
a label for the x axis.}

\item{ylab}{[\code{character(1)}] \cr
a label for the y axis.}

\item{clab}{[\code{character(1)}] \cr
a label for the fill legend.}

\item{droplevels}{[\code{logical(1)}] \cr
drop unused levels from \sQuote{ageGroup} and \sQuote{sex} before plotting.}

\item{drop.NA}{[\code{logical(1)}] \cr
drop rows with missing or \sQuote{NA} values before plotting.}

\item{...}{[] \cr
additional arguments. Ignored.}
}
\description{
Visualize new daily cases
}
