% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSimPara.R
\name{checkSimPara}
\alias{checkSimPara}
\title{checkSimPara}
\usage{
checkSimPara(para = babsimHospitalPara())
}
\arguments{
\item{para}{list: optimization parameters, e.g., generated via \code{\link{babsimHospitalPara}}}
}
\value{
corrected parameter list
}
\description{
check (and correct) parameter list
}
\examples{

x0 <- babsimHospitalPara()
x <- checkSimPara(x0)
}
