% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainTestObjFun.R
\name{checkTestConfig}
\alias{checkTestConfig}
\title{Check test config}
\usage{
checkTestConfig(testConf, TestSimStartDate, TestFieldStartDate, verbosity)
}
\arguments{
\item{testConf}{test configuration}

\item{TestSimStartDate}{test data: sim start date}

\item{TestFieldStartDate}{test data: field start date}

\item{verbosity}{default 0.}
}
\description{
Check test config
}
