% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainTestObjFun.R
\name{checkTestData}
\alias{checkTestData}
\title{Check Test Data}
\usage{
checkTestData(testData, TestSimStartDate, TestFieldStartDate, verbosity)
}
\arguments{
\item{testData}{list of test data: field and sim}

\item{TestSimStartDate}{test data: sim start date}

\item{TestFieldStartDate}{test data: field start date}

\item{verbosity}{default 0.}
}
\description{
Check test data
}
