% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainTestObjFun.R
\name{checkTrainConfig}
\alias{checkTrainConfig}
\title{Check train config}
\usage{
checkTrainConfig(
  trainConf,
  TrainSimStartDate,
  TrainFieldStartDate,
  verbosity = 0
)
}
\arguments{
\item{trainConf}{trainConfig}

\item{TrainSimStartDate}{Start data simulation train data}

\item{TrainFieldStartDate}{Start data simulation field data}

\item{verbosity}{Default 0.}
}
\description{
Check consistency of the training configuration
}
