% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureRangeOpen.R
\name{ensureRangeOpen}
\alias{ensureRangeOpen}
\title{ensureRangeOpen}
\usage{
ensureRangeOpen(x, a, b)
}
\arguments{
\item{x}{value}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
corrected value
}
\description{
Ensure that value belongs to the open interval ]a,b[
}
\examples{
# return 1:
ensureRangeOpen(x = 10, a = 0, b = 1)
# return 0:
ensureRangeOpen(x = 0, a = 0, b = 1)
# return 0.5:
ensureRangeOpen(x = 0.5, a = 0, b = 1)
}
