% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSimulationResults.R
\name{extractSimulationResults}
\alias{extractSimulationResults}
\title{Extract and summarize BaBSim results}
\usage{
extractSimulationResults(envs, conf)
}
\arguments{
\item{envs}{List of simulation environments}

\item{conf}{Configuration for simulation runs in \code{envs}}
}
\value{
A \code{data.table} with columns
\describe{
  \item{\code{date}}{Day on which the resource is required.}
  \item{\code{resource}}{The resource required.}
  \item{\code{replication}}{From which replication the result stems.}
  \item{\code{count}}{Number of units of \code{resource} in use.}
}
}
\description{
Given a list of result environments returned by \code{\link{babsimHospital}}, extract and summarize resource usage for each day.
If the resource usage fluctuates during the day, the maximum usage is reported.
}
