% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BabsimRkiCases.R
\name{fetchRkiCases}
\alias{fetchRkiCases}
\title{Download German daily case summary from RKI}
\usage{
fetchRkiCases(dir = tempdir(), force = FALSE)
}
\arguments{
\item{dir}{[\code{character(1)}] \cr 
directory where downloaded data is stored. Defaults to \code{tempdir()}.}

\item{force}{[\code{bool(1)}] \cr
if true, force download of data even if a previous download is found in \code{dir}.}
}
\value{
An object of class \sQuote{BabsimCasesRki}.
}
\description{
Download German daily case summary from RKI
}
\seealso{
\itemize{
  \item{\code{\link{GermanStates}} for a list of German states and their state id.}
  \item{\code{\link{GermanCounties}} for a list of German counties and theirs county id.}
}
}
