% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funWrapOptimizeSim.R
\name{funWrapOptimizeSim}
\alias{funWrapOptimizeSim}
\title{funWrapOptimizeSim}
\usage{
funWrapOptimizeSim(x, conf, data)
}
\arguments{
\item{x}{num: real values. Will be interpreted as parameter values from \code{babsimHospital}.
Names of these parameters can be obtained via \code{\link{getParameterName}}.}

\item{conf}{list with the following entries:
\describe{
\item{\code{seed}}{seed. Default: 123}
\item{\code{simRepeats}}{simmer repeats}
\item{\code{parallel}}{simmer parallel runs. Default: FALSE}
\item{\code{perCores}}{percentage of cores used for parallel simmer simulations. Default: 0.5 (=50 percent)}
\item{\code{ICU}}{use ICU infection data. Default: FALSE}
\item{\code{logLevel}}{log leved (0 or 1). Default: 0 (no output)}
}
For example: \code{conf <- babsimToolsConf()}}

\item{data}{list with simData and fieldData, e.g. \code{data <- getObkData()}.}
}
\value{
This function returns a  num [1, 1] matrix,
that represents the combined rmse from the three beds.
}
\description{
Wrapper function for \code{funOptimizeSim}
}
\examples{
para <- getStartParameter()
conf <- babsimToolsConf()
data <- getObkData()
funWrapOptimizeSim(x = para, conf = conf, data = data)
}
