% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageData.R
\docType{data}
\name{GermanStates}
\alias{GermanStates}
\title{List of German states}
\format{
A \code{\link[data.table]{data.table}} with 16 rows an 2 variables:
\describe{
  \item{\code{stateId}}{unique id, first two digits of the German AGS.}
  \item{\code{state}}{name of state in German.}
}
}
\source{
Gemeindeverzeichnis-Informationssystem (GV-ISys) of the German Federal Statistics Office
}
\usage{
GermanStates
}
\description{
A dataset containing all 16 German states.
}
\examples{
head(GermanStates)

}
\seealso{
\code{\link{GermanCounties}} for German counties.
}
\keyword{datasets}
