% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainTestObjFun.R
\name{getAndCheckTrainData}
\alias{getAndCheckTrainData}
\title{Get and Check Train Data}
\usage{
getAndCheckTrainData(
  simData,
  fieldData,
  tryOnTestSet = FALSE,
  TrainFieldStartDate,
  TestFieldStartDate,
  verbosity = 0
)
}
\arguments{
\item{simData}{simulation data, i.e., data used for the simulator, e.g. infections.}

\item{fieldData}{field data, i.e., real-world data.}

\item{tryOnTestSet}{logical, default FALSE.}

\item{TrainFieldStartDate}{Start date of the train data set.}

\item{TestFieldStartDate}{Start date of the test data set.}

\item{verbosity}{Default 0.
Only needed if \code{tryOnTestSet} is \code{TRUE}.}
}
\value{
traindata, i.e., list of sim data and field data, checked train data set
}
\description{
Generate and check train data (start and end dates) from field and simulation
data.
}
