% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBounds.R
\name{getBounds}
\alias{getBounds}
\title{getBounds}
\usage{
getBounds()
}
\value{
This function returns a list of two vectors
}
\description{
Returns parameter bounds for babsim runs
(settings version > v10.4.8)
}
\examples{
bounds <- getBounds()
lower <- bounds$lower
upper <- bounds$upper
}
