% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiviLinks.R
\name{getDiviLinks}
\alias{getDiviLinks}
\title{Retrieve links to all DIVI day reports.}
\usage{
getDiviLinks()
}
\value{
List of download URLs to the daily reports.
}
\description{
Retrieve links to all DIVI day reports.
}
