% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimDivi.R
\name{getIcuBeds}
\alias{getIcuBeds}
\title{getIcuBeds}
\usage{
getIcuBeds(data = babsim.hospital::icudata)
}
\arguments{
\item{data}{data.frame with obs. of  9 variables}
}
\value{
data frame with observations of 3 variables:
\describe{
  \item{intensiveBed}{int COVID-19 ICU beds without ventilation}
  \item{intensiveBedVentilation}{int  COVID-19 ICU beds with ventilation}
  \item{Day}{Date, format: "2020-05-01" "2020-05-02" "2020-05-03" "2020-05-04" ...}
 }
}
\description{
Convert the 9 dim DIVI ICU data (bundesland,gemeindeschluessel,..., daten_stand)
into a data.frame with bed, intensiveBedVentilation, and Day
}
\examples{

IcuBeds <- getIcuBeds(data = icudata) 

}
