% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrixD.R
\name{getMatrixD}
\alias{getMatrixD}
\title{getMatrixD}
\usage{
getMatrixD(para = babsimHospitalPara())
}
\arguments{
\item{para}{parameter vector, e.g., generated via
\code{\link{babsimHospitalPara}}. Default: \code{\link{babsimHospitalPara}}}
}
\value{
a matrix with transition durations
}
\description{
Builds the duration matrix
}
\examples{

getMatrixD()
}
