% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimTools.R
\name{getMatrixP}
\alias{getMatrixP}
\title{getMatrixP}
\usage{
getMatrixP(para = babsimHospitalPara())
}
\arguments{
\item{para}{parameter vector, e.g., generated via 
\code{\link{babsimHospitalPara}}. Default: \code{\link{babsimHospitalPara}}}
}
\value{
a matrix with transition probabilities
}
\description{
Builds the probability matrix
}
\examples{

getMatrixP()

}
