% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimTools.R
\name{getObkData}
\alias{getObkData}
\title{getObkData}
\usage{
getObkData(data = babsim.hospital::dataCovidBeds20200624)
}
\arguments{
\item{data}{OBK data. Default: \code{\link{dataCovidBeds20200624}}}
}
\value{
list with simData and fieldData
}
\description{
Generate simData and fieldData from OBK data
}
\examples{

data <- getObkData() 

}
