% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimPara.R
\name{getParameterDataFrame}
\alias{getParameterDataFrame}
\title{getParameterDataFrame}
\usage{
getParameterDataFrame(
  paraList = list(obk = babsim.hospital::obkpara, koeln = babsim.hospital::koelnpara,
    nrw = babsim.hospital::nrwpara)
)
}
\arguments{
\item{paraList}{list of parameter values. 
Each list element has the form \code{obk=obkparam}.
An example list looks as follows: \code{list(obk=obkpara, 
                                            koeln=koelnpara, 
                                            nrw=nrwpara)}}
}
\value{
data.frame with parameters in each column.
}
\description{
Get parameterss (probabilities and durations) of the
babsim.hospital simulator
}
\examples{

df <- getParameterDataFrame(paraList = list(obk=obkpara, 
                                            koeln=koelnpara, 
                                            nrw=nrwpara))
}
