% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimPara.R
\name{getParameterName}
\alias{getParameterName}
\title{getParameterName}
\usage{
getParameterName(n)
}
\arguments{
\item{n}{int: position}
}
\value{
This function returns a  character value, which represents
the name of the n-th \code{x} variable.
}
\description{
Returns the name (chr) of the babsim \code{x} parameter vector.
}
\examples{
getParameterName(16)
 
}
