% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameterNameList.R
\name{getParameterNameList}
\alias{getParameterNameList}
\title{getParameterNameList}
\usage{
getParameterNameList(x)
}
\arguments{
\item{x}{vector of int: positions}
}
\value{
This function returns a vector. Its elements represent
the names of the n-th \code{x} variables.
}
\description{
Returns the names (chr) of the babsim \code{x} parameter vector.
}
\examples{
getParameterNameList(c(16, 18))
}
