% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParaSet.R
\name{getParaSet}
\alias{getParaSet}
\title{getParaSet}
\usage{
getParaSet(region)
}
\arguments{
\item{region}{integer, the region id}
}
\value{
data.frame parameters of that specific region
}
\description{
Load a specific parameter set by region.
}
\examples{
getParaSet(5315)
}
