% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeakVec.R
\name{getPeakVec}
\alias{getPeakVec}
\title{getPeakVec}
\usage{
getPeakVec(
  peakData = c(10, 100),
  StartDate = "2020-03-03",
  EndDate = "2020-06-24"
)
}
\arguments{
\item{peakData}{Vector of time steps and peak heights.}

\item{StartDate}{Day, simulation starts.}

\item{EndDate}{Day, simulation ends.}
}
\value{
This function returns a vector of peaks data.
}
\description{
Generate peak values.
}
\examples{
getPeakVec()
}
