% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionIcu.R
\name{getRegionIcu}
\alias{getRegionIcu}
\title{getRegionIcu Auswahl der ICU Daten fuer eine Region}
\usage{
getRegionIcu(data = babsim.hospital::icudata, region = 5315)
}
\arguments{
\item{data}{Daten, z.B. icudata}

\item{region}{Id der Region, \code{0} Deutschland}
}
\description{
Auswahl anhand der Bundeslaender, Landkreis IDs
}
\examples{
data <- getRegionIcu(babsim.hospital::icudata, 5315)
}
