% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimRki.R
\name{getRegionRki}
\alias{getRegionRki}
\title{getRegionRki Auswahl der RKI Daten fuer eine Region}
\usage{
getRegionRki(data = babsim.hospital::rkidata, region)
}
\arguments{
\item{data}{Daten, z.B. rkidata}

\item{region}{Id der Region, \code{0} Deutschland}
}
\description{
Auswahl anhand der Bundeslaender, Landkreis IDs
}
\examples{
data <- getRegionRki(data = babsim.hospital::rkidata[1:1000,], 
                            region = 0)

}
