% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainTestObjFun.R
\name{getTestData}
\alias{getTestData}
\title{Get Test Data}
\usage{
getTestData(fieldData, simData, TestFieldStartDate, TestSimStartDate)
}
\arguments{
\item{fieldData}{field data}

\item{simData}{sim data}

\item{TestFieldStartDate}{start date}

\item{TestSimStartDate}{start date}
}
\value{
testData test data
}
\description{
Generate test data for final evaluation of one optimization run
}
