% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimvisualize.R
\name{ggVisualizeRki}
\alias{ggVisualizeRki}
\title{ggVisualizeRki ggPlot Visualisierung der RKI Daten}
\usage{
ggVisualizeRki(
  data = babsim.hospital::rkidata,
  region = 5374,
  StartDate = "2020-05-01"
)
}
\arguments{
\item{data}{rki data, e.g., \code{\link{rkidata}}}

\item{region}{Landkreis Id, e.g., \code{5374} oder Bundesland ID, e.g., \code{5}.}

\item{StartDate}{Start (Tag), e.g., \code{"2020-05-01"}}
}
\description{
Quelle: RKI Daten bundesweit
}
\seealso{
\code{\link{rkidata}}
}
