% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVisualizeRkiExtended.R
\name{ggVisualizeRkiExtended}
\alias{ggVisualizeRkiExtended}
\title{ggVisualizeRkiExtended Visualisation of the extended RKI Data}
\usage{
ggVisualizeRkiExtended(
  data = extendRki(),
  region = 5374,
  StartDate = "2020-10-01",
  simplify = TRUE,
  preloadedData = NULL
)
}
\arguments{
\item{data}{rki data as preprocessed by \code{\link{extendRki}}}

\item{region}{Landkreis Id, e.g., \code{5374} oder Bundesland ID, e.g., \code{5}.}

\item{StartDate}{Start (Tag), e.g., \code{'2020-05-01'}}

\item{simplify}{logical. Simplify presentation, Default: TRUE.}

\item{preloadedData}{optional way to pass the result of preloading.}
}
\description{
ggplot RKI data as result from \code{\link{extendRki}}
}
\seealso{
\code{\link{rkidata}}
}
