% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAgegroupToAge.R
\name{mapAgeGroupToAge}
\alias{mapAgeGroupToAge}
\title{mapAgeGroupToAge}
\usage{
mapAgeGroupToAge(x)
}
\arguments{
\item{x}{age vector}
}
\value{
age as a real value
}
\description{
Calculate real valued age based on RKI age classes
}
