% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAgeToAgeGroup.R
\name{mapAgeToAgeGroup}
\alias{mapAgeToAgeGroup}
\title{mapAgeToAgeGroup}
\usage{
mapAgeToAgeGroup(x)
}
\arguments{
\item{x}{age vector}
}
\value{
age class
}
\description{
Calculate age based on RKI age classes
}
