% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimPara.R
\name{mapPToPara}
\alias{mapPToPara}
\title{mapPToPara}
\usage{
mapPToPara(P = getMatrixP(), para)
}
\arguments{
\item{P}{(num) nxn-dim matrix. Values will be mapped onto the probabilities in
\code{babsimHospitalPara}.
Names of these parameters can be obtained via \code{\link{getParameterName}}.}

\item{para}{Parameter list, e.g., generated via \code{\link{babsimHospitalPara}}}
}
\value{
This function returns a parameter list.
}
\description{
\code{mapPToPara} accepts
a nxn matrix. Its values will be mapped onto 
the probability entries of a \code{\link{babsimHospitalPara}}
list.
}
\examples{
para <- babsimHospitalPara()
P <- getMatrixP() 
para <-  mapPToPara(P = P,
                    para = para)
 
}
