% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageDateRange.R
\name{messageDateRange}
\alias{messageDateRange}
\title{Display date range of a vector}
\usage{
messageDateRange(prefix, d)
}
\arguments{
\item{prefix}{string to print before outputting range.}

\item{d}{vector of values.}
}
\value{
Nothing, called for the side effect.
}
\description{
Utility function to display the range of dates in a vector.
}
