% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageData.R
\docType{data}
\name{paras}
\alias{paras}
\title{paras data}
\format{
'data.frame' obs. of  31 variables:
\describe{
  \item{y}{num  311 158 180 232 297 ...}
  \item{x.1}{num  10.55 17.91 3.19 9.5 17.71 ...}
  \item{...}{ ...}
  \item{x.29}{num  1.072 1.015 1.044 1.057 0.556 ...}
  \item{region}{num 5315}
}
}
\usage{
paras
}
\description{
Data: Parameters for all regions, generated through SPOT optimization
}
\details{
Result from the \code{\link{runoptDirect}} run.
}
\keyword{datasets}
