% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BabsimRkiCases.R
\name{rkiGeschlechtToSex}
\alias{rkiGeschlechtToSex}
\title{Map \code{Geschlecht} to biological sex}
\usage{
rkiGeschlechtToSex(geschlecht)
}
\arguments{
\item{geschlecht}{[\code{character(n)}] \cr
character vector with values in \sQuote{M}, \sQuote{W}, and \sQuote{unbekannt}.}
}
\value{
A factor vector with \code{n} elements and levels \sQuote{male} or \sQuote{female}. 
  \sQuote{unbekannt} or other values are mapped to \code{NA}.
}
\description{
Map \code{Geschlecht} to biological sex
}
