% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgamma.R
\name{rtgamma}
\alias{rtgamma}
\title{rtgamma}
\usage{
rtgamma(n = 1, shape = 1, rate = 1, shift = 0, alpha = 0.95)
}
\arguments{
\item{n}{number of observations}

\item{shape}{Gamma shape parameter}

\item{rate}{Gamma rate parameter}

\item{shift}{shift parameter.}

\item{alpha}{upper quantile of gamma distribution. All values above alpha are truncated.}
}
\value{
rtgamma generates random deviates. The length of the result is determined by \code{n}.
}
\description{
Random generation for the shifted and truncated Gamma distribution with parameters
  \code{shape} and \code{scale}.
}
\examples{

rtgamma(n = 1, shape = 1, rate = 1, shift = 1, alpha = 0.95)
}
