% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip_if_quicktest.R
\name{skip_if_quicktest}
\alias{skip_if_quicktest}
\title{Skip test if in quicktest mode}
\usage{
skip_if_quicktest()
}
\description{
Skip test if option \sQuote{babsim.hospital.quicktest} is \code{TRUE} or if
the environment variable \sQuote{BABSIM_HOSPTIAL_QUICKTEST} is set.
}
