% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimRki.R
\name{switchRkiMeldedatumToRefdatum}
\alias{switchRkiMeldedatumToRefdatum}
\title{switchRkiMeldedatumToRefdatum Change RKI data File}
\usage{
switchRkiMeldedatumToRefdatum(data = rkidata, overwrite = TRUE)
}
\arguments{
\item{data}{RKI Data}

\item{overwrite}{logical Overwrite existing file. Default \code{TRUE}.}
}
\description{
Use Refdatum instead of Meldedatum in
\code{\link{rkidata}}. This is the original configuration provides by RKI.
It is recommended to run
install and restart the package afterwards.
}
