% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchRkiRefdatumToMeldedatum.R
\name{switchRkiRefdatumToMeldedatum}
\alias{switchRkiRefdatumToMeldedatum}
\title{switchRkiRefdatumToMeldedatum Change RKI data File}
\usage{
switchRkiRefdatumToMeldedatum(rkidata = rkidata, overwrite = TRUE)
}
\arguments{
\item{rkidata}{RKI Data}

\item{overwrite}{logical Overwrite existing file. Default \code{TRUE}.}
}
\description{
Use Meldedatum instead of Refdatum in
\code{\link{rkidata}}. It is recommended to run
install and restart the package afterwards.
}
