% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukpara.R
\docType{data}
\name{ukpara}
\alias{ukpara}
\title{ukpara data}
\format{
'data.frame':	 obs. of  29 variables:
\describe{
  \item{y}{num  311 158 180 232 297 ...}
  \item{x.1}{num  10.55 17.91 3.19 9.5 17.71 ...}
  \item{...}{ ...}
  \item{x.29}{num  1.072 1.015 1.044 1.057 0.556 ...}
}
}
\usage{
ukpara
}
\description{
Data: ukpara para generated with SPOT.
}
\details{
Result from the \code{runoptUK} run.
Use \code{yx <- ukpara[ukpara$y == min(ukpara$y), ]
x <- yx[1,2:dim(ukpara)[2]]
} 
to extract the best parameter set x.
}
\keyword{datasets}
