% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimDivi.R
\name{updateIcudataFile}
\alias{updateIcudataFile}
\title{updateIcudataFile Update ICU data File}
\usage{
updateIcudataFile(oldData = babsim.hospital::icudataFull, overwrite = TRUE)
}
\arguments{
\item{oldData}{Old icu (DIVI) data, default \code{babsim.hospital::icudata}.}

\item{overwrite}{logical Overwrite existing file. Default \code{TRUE}.}
}
\value{
True if new data was downloaded, otherwise false
}
\description{
Update icudata (divi)
}
