% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimTools.R
\name{updateMatrixP}
\alias{updateMatrixP}
\title{updateMatrixP}
\usage{
updateMatrixP(P = getMatrixP(), u)
}
\arguments{
\item{P}{matrix P. Default \code{\link{getMatrixP}}}

\item{u}{list of factors used for the update}
}
\value{
a matrix with updated transition probabilities
}
\description{
Updates the probability matrix
}
\examples{

u <- list(k=2)
R <- updateMatrixP(u=u)

}
