% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateParaSet.R
\name{updateParaSet}
\alias{updateParaSet}
\title{updateParaSet}
\usage{
updateParaSet(paramDF, region, path = NULL)
}
\arguments{
\item{paramDF}{the data frame with the new parameters that should replace the old ones}

\item{region}{integer, the region id}

\item{path}{path to the old para file.}
}
\value{
parameter set
}
\description{
Delete old parameters from paras and replace them with new ones.
}
\examples{
getParaSet(5315)
}
