% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageData.R
\docType{data}
\name{vaccineCounts}
\alias{vaccineCounts}
\title{Amount of vaccinated people in germany}
\format{
'data.frame':  obs. of  4 variables:
\describe{
  \item{vaccinated}{num  311 158 180 232 297 ...}
  \item{quote}{num  10.55 17.91 3.19 9.5 17.71 ...}
  \item{state}{char Berlin Bayern ...}
  \item{date}{date  2020-12-29 2020-12-30 ...}
}
}
\usage{
vaccineCounts
}
\description{
Data: Amount of vaccinated people in germany. Collected by day and state.
Additionally the vaccine-quota (vaccinated/capita) is stored
}
\keyword{datasets}
