% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babsimvisualize.R
\name{visualizeGraph}
\alias{visualizeGraph}
\title{visualizeGraph Visualisierung der Wahrscheinlichkeiten und Dauern}
\usage{
visualizeGraph(para = babsimHospitalPara(), option = "P")
}
\arguments{
\item{para}{parameter}

\item{option}{Option: plot probabilities ("P") or durations ("D")}
}
\description{
Ergebnisse
}
\examples{

 para = babsimHospitalPara()
 visualizeGraph(para = para , option = "P")

}
