% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeRkiEvents.R
\name{visualizeRkiEvents}
\alias{visualizeRkiEvents}
\title{visualizeRkiEvents Visualisation of the pre-processed RKI Data}
\usage{
visualizeRkiEvents(
  data = getRkiData(babsim.hospital::rkidata),
  region = 0,
  StartDate = "2020-05-01"
)
}
\arguments{
\item{data}{rki data as preprocessed by \code{\link{getRkiData}}}

\item{region}{Landkreis Id, e.g., \code{5374} oder Bundesland ID, e.g., \code{5}.}

\item{StartDate}{Start (Tag), e.g., \code{'2020-05-01'}}
}
\description{
RKI data as result from \code{getRkiData(babsim.hospital::rkidata)}
}
\examples{

p <- visualizeRkiEvents(getRkiData(babsim.hospital::rkidata[1:1000, ]))
}
\seealso{
\code{\link{getRkiData}}
}
