% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeUK.R
\name{visualizeUK}
\alias{visualizeUK}
\title{visualizeUK}
\usage{
visualizeUK(data, region = 5315)
}
\arguments{
\item{data}{Date}

\item{region}{Landkreis Id, e.g., \code{5374} fuer OBK, \code{5315} fuer Koeln,
\code{0} fuer Deutschland,
oder Bundesland ID, e.g., \code{5} fuer NRW.}
}
\description{
Anzeigen der UK Daten
}
\examples{

\dontrun{
res <- visualizeUK()
}

}
