% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameterDataFrame.R
\name{getParameterDataFrame}
\alias{getParameterDataFrame}
\title{getParameterDataFrame}
\usage{
getParameterDataFrame(
  paraList = list(obk = getParaSet(5374), koeln = getParaSet(5315))
)
}
\arguments{
\item{paraList}{list of parameter values.
Each list element has the form \code{obk=getParaSet(5374)}.}
}
\value{
data.frame with parameters in each column.
}
\description{
Get parameterss (probabilities and durations) of the
babsim.hospital simulator
}
\examples{

df <- getParameterDataFrame()
}
