% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateIcudataFile.R
\name{updateIcudataFile}
\alias{updateIcudataFile}
\title{updateIcudataFile Update ICU data File}
\usage{
updateIcudataFile(overwrite = TRUE)
}
\arguments{
\item{overwrite}{logical Overwrite existing file. Default \code{TRUE}.}
}
\value{
True if new data was downloaded, otherwise false
}
\description{
Update icudata (divi)
}
